Send serialzied notes to print view
