+++ src/selectors/items.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const { createSelector: memo } = require('reselect')
-const { getMetadataFields } = require('./metadata')
 const { pluck } = require('./util')
 const EMPTY = {}
 
@@ -31,46 +30,9 @@ const getSelectedItemTemplate = memo(
   }
 )
 
-const expandPhoto = (photo, metadata, ontology) => (
-  (photo == null) ? null : {
-    ...photo,
-    data: getMetadataFields(null, {
-      compact: true,
-      data: metadata[photo.id],
-      props: ontology.props,
-      template: ontology.template[photo.template]
-    })
-  }
-)
-
-const expandItem = (item, photos, metadata, ontology) => (
-  (item == null) ? null : {
-    ...item,
-    data: getMetadataFields(null, {
-      compact: true,
-      data: metadata[item.id],
-      props: ontology.props,
-      template: ontology.template[item.template]
-    }),
-    photos: item.photos.map(id =>
-      expandPhoto(photos[id], metadata, ontology))
-  }
-)
-
-const getPrintableItems = memo(
-  ({ nav, qr }) => (nav.items.length > 0 ? nav.items : qr.items),
-  ({ items }) => items,
-  ({ photos }) => photos,
-  ({ metadata }) => metadata,
-  ({ ontology }) => ontology,
-  (ids, items, photos, metadata, ontology) =>
-    ids.map(id => expandItem(items[id], photos, metadata, ontology))
-)
-
 module.exports = {
   getItems,
   getListHold,
-  getPrintableItems,
   getSelectedItems,
   getSelectedItemTemplate,
   getVisibleItems
